Extract report reducer
